requirements.txt
openai
chromadb
scrapy
pandas
fastapi
websockets
uvicorn[standard]
pydantic
python-dotenv